+++ openbrokerapi/api.py
@@ -197,6 +197,10 @@ def get_blueprint(service_broker: ServiceBroker,
                 error="AsyncRequired",
                 description="This service plan requires client support for asynchronous service operations."
             )), HTTPStatus.UNPROCESSABLE_ENTITY
+        except errors.ErrConcurrentInstanceAccess:
+            error_response = ErrorResponse(error='ConcurrencyError',
+                                           description='The Service Broker does not support concurrent requests that mutate the same resource.')
+            return to_json_response(error_response), HTTPStatus.UNPROCESSABLE_ENTITY
 
         if result.is_async:
             return to_json_response(UpdateResponse(result.operation, result.dashboard_url)), HTTPStatus.ACCEPTED
@@ -229,6 +233,10 @@ def get_blueprint(service_broker: ServiceBroker,
                 error="RequiresApp",
                 description="This service supports generation of credentials through binding an application only."
             )), HTTPStatus.UNPROCESSABLE_ENTITY
+        except errors.ErrConcurrentInstanceAccess:
+            error_response = ErrorResponse(error='ConcurrencyError',
+                                           description='The Service Broker does not support concurrent requests that mutate the same resource.')
+            return to_json_response(error_response), HTTPStatus.UNPROCESSABLE_ENTITY
 
         response = BindResponse(
             credentials=result.credentials,
@@ -267,6 +275,10 @@ def get_blueprint(service_broker: ServiceBroker,
         except errors.ErrBindingDoesNotExist as e:
             logger.exception(e)
             return to_json_response(EmptyResponse()), HTTPStatus.GONE
+        except errors.ErrConcurrentInstanceAccess:
+            error_response = ErrorResponse(error='ConcurrencyError',
+                                           description='The Service Broker does not support concurrent requests that mutate the same resource.')
+            return to_json_response(error_response), HTTPStatus.UNPROCESSABLE_ENTITY
 
         if result.is_async:
             return to_json_response(UnbindResponse(result.operation)), HTTPStatus.ACCEPTED
@@ -300,6 +312,10 @@ def get_blueprint(service_broker: ServiceBroker,
                 error="AsyncRequired",
                 description="This service plan requires client support for asynchronous service operations."
             )), HTTPStatus.UNPROCESSABLE_ENTITY
+        except errors.ErrConcurrentInstanceAccess:
+            error_response = ErrorResponse(error='ConcurrencyError',
+                                           description='The Service Broker does not support concurrent requests that mutate the same resource.')
+            return to_json_response(error_response), HTTPStatus.UNPROCESSABLE_ENTITY
 
         if result.is_async:
             return to_json_response(DeprovisionResponse(result.operation)), HTTPStatus.ACCEPTED