+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 Changelog
 =============
 
+  - ErrConcurrentInstanceAccess exception on an instance update, an instance deprovision, a binding or unbinding must return 422 not 500.
+  - ErrPlanChangeNotSupported exception must return 400 not 500.
+
+
 **v4.1**
   - Drop Python 3.5 support
   - Fix ´bind´ in multi broker setup (#117); thx @vaxvms